+++ inc/BoundingBox.h
@@ -31,19 +31,19 @@ namespace degate
 	class BoundingBox
 	{
 	private:
-		int min_x, max_x, min_y, max_y;
+        float min_x, max_x, min_y, max_y;
 
 	public:
 		BoundingBox();
 
-		BoundingBox(int min_x, int max_x, int min_y, int max_y);
-		BoundingBox(int width, int height);
+		BoundingBox(float min_x, float max_x, float min_y, float max_y);
+		BoundingBox(float width, float height);
 		BoundingBox(const BoundingBox&);
 
 		virtual ~BoundingBox();
 
 		BoundingBox const& get_bounding_box() const;
-		bool in_shape(int x, int y, int max_distance = 0) const;
+		bool in_shape(float x, float y, float max_distance = 0) const;
 
 		/**
 		 * Check if this bounding box is completly within the bounding box given as parameter.
@@ -56,27 +56,27 @@ namespace degate
 		bool intersects(BoundingBox const& rect) const;
 		bool complete_within(BoundingBox const& rect) const;
 
-		unsigned int get_width() const;
-		unsigned int get_height() const;
+        float get_width() const;
+        float get_height() const;
 
-		int get_min_x() const;
-		int get_max_x() const;
-		int get_min_y() const;
-		int get_max_y() const;
+        float get_min_x() const;
+        float get_max_x() const;
+        float get_min_y() const;
+        float get_max_y() const;
 
-		int get_center_x() const;
-		int get_center_y() const;
+        float get_center_x() const;
+        float get_center_y() const;
 
-		void set_min_x(int min_x);
-		void set_min_y(int min_y);
-		void set_max_x(int max_x);
-		void set_max_y(int max_y);
+		void set_min_x(float min_x);
+		void set_min_y(float min_y);
+		void set_max_x(float max_x);
+		void set_max_y(float max_y);
 
-		void set(int min_x, int max_x, int min_y, int max_y);
+		void set(float min_x, float max_x, float min_y, float max_y);
 
-		void shift_x(int delta_x);
-		void shift_y(int delta_y);
-		void shift(int delta_x, int delta_y);
+		void shift_x(float delta_x);
+		void shift_y(float delta_y);
+		void shift(float delta_x, float delta_y);
 
 		void print(std::ostream& os = std::cout, int n_tabs = 0) const;
 		std::string to_string() const;
